requirements.txt
aiogram
aiohttp
pydnevnikruapi
sqlalchemy[asyncio]
asyncpg
aiosqlite
python-dotenv